backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
firebase-admin==6.2.0
pydantic==2.6.1
pytest==7.4.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
redis==5.0.1
pytest-asyncio==0.23.5
google-cloud-firestore==2.11.1
google-cloud-storage==2.14.0
email-validator==2.1.0.post1
aiofiles==23.2.1
reportlab==4.0.7
pytest-cov==4.1.0
pytest-env==1.1.3
pytest-mock>=3.14.0,<4.0.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
typing-extensions>=4.13.2,<5.0.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
weasyprint==65.1
jinja2==3.1.2
httpx==0.27.0
psycopg2-binary